# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LifeTrigger Engine API
  version: 1.0.0
  description: >
    API determinística para cálculo de necessidade de Seguro de Vida, Score de Proteção e Gap.
    Não utiliza Inteligência Artificial nas regras centrais.
    Totalmente stateless, auditável e segura.
servers:
  - url: https://api.lifetrigger.example.com/api/v1
    description: Production Server
  - url: http://localhost:3000/api/v1
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticação.
  schemas:
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Os dados fornecidos não são válidos."
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "renda_mensal"
              message:
                type: string
                example: "A renda mensal é obrigatória e deve ser maior que zero."
    EvaluationRequest:
      type: object
      required:
        - personalContext
        - financialContext
        - familyContext
        - operationalData
      properties:
        personalContext:
          type: object
          properties:
            age:
              type: integer
              example: 30
            maritalStatus:
              type: string
              enum: [SOLTEIRO, CASADO, DIVORCIADO, VIUVO]
              example: "CASADO"
            professionRiskLevel:
              type: string
              enum: [BAIXO, MEDIO, ALTO, MUITO_ALTO]
              example: "BAIXO"
            isSmoker:
              type: boolean
              example: false
        financialContext:
          type: object
          properties:
            monthlyIncome:
              type: object
              properties:
                exactValue:
                  type: number
                  example: 10000.00
            emergencyFundMonths:
              type: integer
              example: 2
            debts:
              type: object
              properties:
                totalAmount:
                  type: number
                  example: 50000.00
                remainingTermMonths:
                  type: integer
                  example: 36
            currentLifeInsurance:
              type: object
              properties:
                coverageAmount:
                  type: number
                  example: 250000.00
                policyType:
                  type: string
                  enum: [TEMPORARIO, VIDA_INTEIRA, RESGATAVEL, GRUPO_EMPRESARIAL, ACIDENTES_PESSOAIS, PRESTAMISTA]
                  example: "TEMPORARIO"
            educationCosts:
              type: object
              description: "Custos de educação dos dependentes (opcional)"
              properties:
                totalEstimatedCost:
                  type: number
                  example: 120000.00
            estate:
              type: object
              description: "Patrimônio para cálculo de ITCMD e inventário (opcional)"
              properties:
                totalEstateValue:
                  type: number
                  example: 800000.00
                state:
                  type: string
                  description: "Sigla UF do estado (alíquota ITCMD varia de 2-8%)"
                  example: "SP"
        familyContext:
          type: object
          properties:
            dependentsCount:
              type: integer
              example: 1
            dependentsAges:
              type: array
              items:
                type: integer
              example: [ 3, 5 ]
        operationalData:
          type: object
          properties:
            originChannel:
              type: string
              example: "APP_MOBILE"
            hasExplicitActiveConsent:
              type: boolean
              example: true
            consentId:
              type: string
              example: "term-123"
            tenantId:
              type: string
              format: uuid
            recentLifeTrigger:
              type: boolean
              example: false

    EvaluationResponse:
      type: object
      properties:
        recommendedCoverageAmount:
          type: number
          format: float
          example: 1340000.00
        currentCoverageAmount:
          type: number
          format: float
          example: 100000.00
        protectionGapAmount:
          type: number
          format: float
          example: 1240000.00
        protectionGapPercentage:
          type: number
          format: float
          example: 92.53
        protectionScore:
          type: integer
          example: 0
        coverageEfficiencyScore:
          type: integer
          example: 100
        riskClassification:
          type: string
          enum: [CRITICO, MODERADO, ADEQUADO]
          example: "CRITICO"
        coverageStatus:
          type: string
          enum: [SUBPROTEGIDO, ADEQUADO, SOBRESEGURADO]
          example: "SUBPROTEGIDO"
        recommendedAction:
          type: string
          enum: [AUMENTAR, MANTER, REDUZIR, REVISAR]
          example: "AUMENTAR"
        incomeReplacementAmount:
          type: number
          description: "Componente de reposição de renda"
        debtClearanceAmount:
          type: number
          description: "Componente de liquidação de dívidas"
        transitionReserveAmount:
          type: number
          description: "Componente de reserva de transição"
        educationCostsAmount:
          type: number
          description: "Componente de custos de educação"
        itcmdCostAmount:
          type: number
          description: "Componente de ITCMD (imposto sucessório)"
        inventoryCostAmount:
          type: number
          description: "Componente de custos de inventário"
        regrasAplicadas:
          type: array
          description: "Usado para UI Icons / Exibição Front-End"
          items:
            type: string
          example: ["RULE_PENALTY_HIGH_DEBT"]
        justificationsStructured:
          type: array
          description: "Usado para integração B2B"
          items:
            type: object
        justificationsRendered:
          type: array
          items:
            type: string
        audit:
          type: object
          description: "Usado para armazenamento Legal/Compliance no Back-End Integrador"
          properties:
            engineVersion:
              type: string
              example: "1.0.0"
            ruleSetVersion:
              type: string
              example: "2026.02"
            appliedRules:
              type: array
              items:
                type: string
            timestamp:
              type: string
              format: date-time
            consentId:
              type: string
              example: "term-123"
        brokerInsights:
          type: array
          description: "Insights acionáveis para o corretor apresentar ao cliente"
          items:
            type: object
            properties:
              icon:
                type: string
                example: "lightbulb"
              title:
                type: string
                example: "Oportunidade de Proteção"
              description:
                type: string
                example: "O cliente tem uma lacuna significativa na cobertura."
              priority:
                type: string
                enum: [high, medium, low]
    TriggerEventRequest:
      type: object
      required:
        - trigger_type
        - base_request
      properties:
        trigger_type:
          type: string
          enum: 
            - RENDA_SUBIU
            - NASCEU_FILHO
            - CASOU
            - DIVORCIO
            - FINANCIAMENTO_NOVO
            - MUDOU_EMPRESA
            - REVISAO_ANUAL
          example: "NASCEU_FILHO"
        description:
          type: string
          example: "Nascimento do segundo filho."
        event_date:
          type: string
          format: date-time
          example: "2026-02-20T00:00:00Z"
        base_request:
          $ref: '#/components/schemas/EvaluationRequest'
    EngineVersionResponse:
      type: object
      properties:
        engine_version:
          type: string
          example: "1.0.0"
        ruleset_version:
          type: string
          example: "2026.02"
        description:
          type: string
          example: "LifeTrigger Engine Deterministic Rules"

security:
  - BearerAuth: []

paths:
  /evaluations:
    post:
      summary: Executar avaliação completa
      description: Processa os dados de contexto do cliente e retorna a recomendação determinística do Seguro de Vida.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationRequest'
      responses:
        '200':
          description: "Avaliação executada com sucesso"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationResponse'
        '400':
          description: "Erro de validação nos dados fornecidos"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Não autorizado (Token JWT inválido ou ausente)"
        '422':
          description: "Entidade não processável (Ex: consentimento_ativo falso)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "CONSENT_REQUIRED"
                message: "O consentimento ativo é obrigatório para realizar a avaliação."
                details: []
        '500':
          description: "Erro interno do servidor"

  /evaluations/{evaluationId}:
    get:
      summary: Recuperar avaliação passada
      description: Recupera o registro imutável de uma avaliação garantindo total reprodutibilidade.
      parameters:
        - name: evaluationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: "Avaliação recuperada com sucesso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  original_request:
                    $ref: '#/components/schemas/EvaluationRequest'
                  result:
                    $ref: '#/components/schemas/EvaluationResponse'
        '401':
          description: "Não autorizado"
        '404':
          description: "Avaliação não encontrada"

  /admin/audit/evaluations/{evaluationId}/verify:
    get:
      summary: Verificação Criptográfica da Avaliação
      description: Rota restrita para auditores recálcular e comparar o Hash SHA-256 de uma requisição salva, garantindo que o algoritmo e o estado financeiro do cliente não sofreram adulterações no log.
      parameters:
        - name: evaluationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: "Verificação de integridade concluída"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [PASS, FAIL]
                  expected_hash:
                    type: string
                    example: "a8b7c9..."
                  actual_hash:
                    type: string
                    example: "a8b7c9..."
        '401':
          description: "Não autorizado (Não é admin)"
        '404':
          description: "Avaliação não encontrada"

  /triggers:
    post:
      summary: Registrar gatilho de vida
      description: Registra um evento de vida (Life Trigger) e retorna uma nova avaliação atualizada, forçando o status de REVISAR e incorporando no histórico.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerEventRequest'
      responses:
        '200':
          description: "Gatilho processado e avaliação secundária gerada"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationResponse'
        '400':
          description: "Erro de validação"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Não autorizado"
        '422':
          description: "Regra de negócio impediu o processamento do gatilho (Ex: consentimento revogado)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "CONSENT_REQUIRED"
                message: "O consentimento ativo é obrigatório para registrar gatilhos."
                details: []
        '500':
          description: "Erro interno do servidor"

  /engine/versions:
    get:
      summary: "Listar versões do motor"
      description: "Retorna a versão do motor de regras e configurações vigentes associadas à avaliação."
      responses:
        '200':
          description: "Versões retornadas com sucesso"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngineVersionResponse'
        '401':
          description: "Não autorizado"
        '500':
          description: "Erro interno do servidor"